Added fog, improved physics, fiexed bugs

+++ Minecraft/Player.h
@@ -19,7 +19,7 @@ public:
 	sf::Vector2f m_camera_angle;
 private:
 	bool m_is_keys_pressed[6] = { false };
-	float m_direction_speed[6] = { 0 };
+	double m_direction_speed[6] = { 0 };
 
 	std::vector<std::pair<World::block_id, int>> m_inventory;
 	World::block_id m_curr_block;
@@ -35,7 +35,7 @@ public:
 	void init(World::Map *map);
 
 	/* calculate movement */
-	void update(float time);
+	void update(double time);
 	void input(sf::Event& e);
 
 	/* eponymous */
